+++ src/gamspy/_symbols/equation.py
@@ -863,7 +863,6 @@ class Equation(gt.Equation, Symbol):
 
         if self._records is not None and self.domain_forwarding:
             self._domainForwarding()
-            self._mark_forwarded_domain_sets(self.domain_forwarding)
 
             # reset state check flags for all symbols in the container
             for symbol in self.container.data.values():